Made fx_renderer private
